+++ src/Door.cpp
@@ -5,7 +5,7 @@
 #include "Door.h"
 
 Game::Door::Door(int posX, int posY) {
-    setPos(posX, posY);
+    setPos((float) posX, (float) posY);
     setCollRec(this->getPos().x, this->getPos().y, 24, 24);
 
 }
